server/requirements.txt
psycopg2==2.9.9
protobuf==3.20.0
sendgrid==6.8.0
google-auth==2.27.0
google_auth_oauthlib==1.2.0
google-api-python-client==2.116.0
testresources==2.0.1
launchpadlib==1.10.13
flask==3.0.3